+++ pycaptions/development/__init__.py
@@ -1,5 +1,4 @@
 from .captionsFormat import CaptionsFormat
-from .block import Block
 from .blockType import BlockType
 from .wrappers import (
     captionsWriter,