+++ pycaptions/development/blocks/blockLayout.py
@@ -1,0 +1,6 @@
+from .blockFormat import BlockFormat
+from ..blockType import BlockType
+
+class LayoutBlock(BlockFormat):
+    def __init__(self, value=None, **options):
+        super().__init__(BlockType.LAYOUT, value, **options)