+++ pycaptions/development/blocks/blockMetadata.py
@@ -1,0 +1,7 @@
+from .blockFormat import BlockFormat
+from ..blockType import BlockType
+
+
+class MetadataBlock(BlockFormat):
+    def __init__(self, value=None, **options):
+        super().__init__(BlockType.METADATA, value, **options)